+++ packages/devtools/lib/src/profiler/cpu_profile_service.dart
@@ -8,6 +8,11 @@ import 'package:vm_service_lib/vm_service_lib.dart';
 import '../globals.dart';
 import '../profiler/cpu_profile_model.dart';
 
+/// Default period at which the VM will collect CPU samples.
+///
+/// This value is applied to the profile_period VM flag.
+const int defaultSamplePeriod = 250;
+
 /// Manages interactions between the Cpu Profiler and the VmService.
 class CpuProfilerService {
   Future<CpuProfileData> getCpuProfile({