+++ packages/devtools/lib/src/ui/service_extension_elements.dart
@@ -1,0 +1,191 @@
+// Copyright 2019 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import '../globals.dart';
+import '../service_extensions.dart';
+import '../service_registrations.dart';
+import '../utils.dart';
+import 'analytics.dart' as ga;
+import 'elements.dart';
+import 'primer.dart';
+
+List<CoreElement> getServiceExtensionElements() {
+  return [
+    div(c: 'btn-group collapsible-1200 nowrap margin-left')
+      ..add(<CoreElement>[
+        ServiceExtensionButton(performanceOverlay).button,
+        ServiceExtensionButton(slowAnimations).button,
+      ]),
+    div(c: 'btn-group collapsible-1200 nowrap margin-left')
+      ..add(<CoreElement>[
+        ServiceExtensionButton(debugPaint).button,
+        ServiceExtensionButton(debugPaintBaselines).button,
+      ]),
+    div(c: 'btn-group collapsible-1400 nowrap margin-left')
+      ..add(<CoreElement>[
+        ServiceExtensionButton(repaintRainbow).button,
+        ServiceExtensionButton(debugAllowBanner).button,
+      ]),
+    div(c: 'btn-group nowrap margin-left')
+      ..add(ServiceExtensionSelector(togglePlatformMode).selector)
+  ];
+}
+
+/// Button that calls a service extension.
+///
+/// Service extensions can be found in [service_extensions.dart].
+class ServiceExtensionButton {
+  ServiceExtensionButton(this.extensionDescription) {
+    button = PButton.icon(
+      extensionDescription.description,
+      extensionDescription.icon,
+      title: extensionDescription.disabledTooltip,
+    )..small();
+
+    final extensionName = extensionDescription.extension;
+
+    // Disable button for unavailable service extensions.
+    button.disabled = !serviceManager.serviceExtensionManager
+        .isServiceExtensionAvailable(extensionName);
+    serviceManager.serviceExtensionManager.hasServiceExtension(
+        extensionName, (available) => button.disabled = !available);
+
+    button.click(() => _click());
+
+    _updateState();
+  }
+
+  final ToggleableServiceExtensionDescription extensionDescription;
+  PButton button;
+
+  void _click() {
+    ga.select(extensionDescription.gaScreenName, extensionDescription.gaItem);
+
+    final bool wasSelected = button.element.classes.contains('selected');
+    serviceManager.serviceExtensionManager.setServiceExtensionState(
+      extensionDescription.extension,
+      !wasSelected,
+      wasSelected
+          ? extensionDescription.disabledValue
+          : extensionDescription.enabledValue,
+    );
+  }
+
+  void _updateState() {
+    // Select button whose state is already enabled.
+    serviceManager.serviceExtensionManager
+        .getServiceExtensionState(extensionDescription.extension, (state) {
+      final extensionEnabled = state.value == extensionDescription.enabledValue;
+      button.toggleClass('selected', extensionEnabled);
+      button.tooltip = extensionEnabled
+          ? extensionDescription.enabledTooltip
+          : extensionDescription.disabledTooltip;
+    });
+  }
+}
+
+/// Button that calls a registered service from flutter_tools. Registered
+/// services can be found in [service_registrations.dart].
+class RegisteredServiceExtensionButton {
+  RegisteredServiceExtensionButton(
+    this.serviceDescription,
+    this.action,
+    this.errorAction,
+  ) {
+    button = PButton.icon(
+      serviceDescription.title,
+      serviceDescription.icon,
+      title: serviceDescription.title,
+    )
+      ..small()
+      ..hidden(true);
+
+    // Only show the button if the device supports the given service.
+    serviceManager.hasRegisteredService(
+      serviceDescription.service,
+      (registered) {
+        button.hidden(!registered);
+      },
+    );
+
+    button.click(() => _click());
+  }
+
+  final RegisteredServiceDescription serviceDescription;
+  final VoidAsyncFunction action;
+  final VoidFunctionWithArg errorAction;
+  PButton button;
+
+  void _click() async {
+    try {
+      button.disabled = true;
+      await action();
+    } catch (e) {
+      errorAction(e);
+    } finally {
+      button.disabled = false;
+    }
+  }
+}
+
+/// Dropdown selector that calls a service extension.
+///
+/// Service extensions can be found in [service_extensions.dart].
+class ServiceExtensionSelector {
+  ServiceExtensionSelector(this.extensionDescription) {
+    selector = PSelect()
+      ..small()
+      ..clazz('button-bar-dropdown')
+      ..change(_handleSelect)
+      ..tooltip = extensionDescription.tooltips.first ??
+          extensionDescription.description;
+
+    extensionDescription.displayValues.forEach(selector.option);
+
+    final extensionName = extensionDescription.extension;
+
+    // Disable selector for unavailable service extensions.
+    selector.disabled = !serviceManager.serviceExtensionManager
+        .isServiceExtensionAvailable(extensionName);
+    serviceManager.serviceExtensionManager.hasServiceExtension(
+        extensionName, (available) => selector.disabled = !available);
+
+    _updateState();
+  }
+
+  final ServiceExtensionDescription extensionDescription;
+
+  PSelect selector;
+
+  String _selectedValue;
+
+  void _handleSelect() {
+    if (selector.value == _selectedValue) return;
+
+    ga.select(extensionDescription.gaScreenName, extensionDescription.gaItem);
+
+    final extensionValue = extensionDescription
+        .values[extensionDescription.displayValues.indexOf(selector.value)];
+
+    serviceManager.serviceExtensionManager.setServiceExtensionState(
+      extensionDescription.extension,
+      true,
+      extensionValue,
+    );
+
+    _selectedValue = selector.value;
+  }
+
+  void _updateState() {
+    // Select option whose state is already enabled.
+    serviceManager.serviceExtensionManager
+        .getServiceExtensionState(extensionDescription.extension, (state) {
+      if (state.value != null) {
+        final selectedIndex = extensionDescription.values.indexOf(state.value);
+        selector.selectedIndex = selectedIndex;
+        _selectedValue = extensionDescription.displayValues[selectedIndex];
+      }
+    });
+  }
+}